Port minipy itself to python3.
This is necessary to support incompatible syntax between python2 and
python3 ASTs.

In addition to general utf-8/bytes/string handling, the two biggest
changes are that arguments are handled different in python3 and
integers, floats, and strings are all treated as Constant rather than
having their own class. A less significant change is that Try is now
a single class rather than two different classes for the TryExcept and
TryFinally cases.

A couple of differences that are fixed yet (expression test cases for
them are commented out in this commit):
- negative numbers in an expression are encoded as a positive number
  within a '-' unary operation.
- explicit unicode strings (starting with a u...) are not yet handled.

Also, remove the testCoding test case file since all python3 files are
UTF-8 now.